Add a WebTask.io script for Quizlet API
The Quizlet API does not use the correct CORS headers which makes it
difficult to call from a browser with default security. In order to
abide with security policy, the webtask calls the quizlet from outside
the browser context and then passes on the results with the correct CORS
headers necessary. A workaround that works